Let there be art!
Added new scene (LodgeMap) complete with all the trimmings. Uploaded a bunch of art from Andres and made a ton of materials to go with that stuff. Made a coffee cup prefab and a player prefab that has the actual ghost, implemented a following camera using the cinemachine package and set up some other camera tools for the future